+++ filebeat/inputsource/common/dgram/server.go
@@ -20,8 +20,6 @@ package dgram
 import (
 	"context"
 	"net"
-	"runtime"
-	"strings"
 	"time"
 
 	"github.com/elastic/go-concert/ctxtool"
@@ -36,15 +34,8 @@ const windowErrBuffer = "A message sent on a datagram socket was larger than the
 	" buffer or some other network limit, or the buffer used to receive a datagram into was smaller" +
 	" than the datagram itself."
 
-// HandlerFactory returns a ConnectionHandler func
-type HandlerFactory func(config ListenerConfig) ConnectionHandler
-
-type ConnectionHandler func(context.Context, net.PacketConn) error
-
+// ListenerFactory is used to craete connections based on the configuration.
 type ListenerFactory func() (net.PacketConn, error)
-
-// MetadataFunc defines callback executed when a line is read from the split handler.
-type MetadataFunc func(net.Conn) inputsource.NetworkMetadata
 
 type ListenerConfig struct {
 	Timeout        time.Duration
@@ -75,67 +66,6 @@ func NewListener(
 		connect:  connect,
 		tg:       unison.TaskGroup{},
 	}
-}
-
-// DatagramReaderFactory allows creation of a handler that has splitting capabilities.
-func DatagramReaderFactory(
-	family inputsource.Family,
-	logger *logp.Logger,
-	callback inputsource.NetworkFunc,
-) HandlerFactory {
-	return func(config ListenerConfig) ConnectionHandler {
-		return ConnectionHandler(func(ctx context.Context, conn net.PacketConn) error {
-			//metadata := metadataCallback(conn)
-			for ctx.Err() == nil {
-
-				buffer := make([]byte, config.MaxMessageSize)
-				conn.SetDeadline(time.Now().Add(config.Timeout))
-
-				// If you are using Windows and you are using a fixed buffer and you get a datagram which
-				// is bigger than the specified size of the buffer, it will return an `err` and the buffer will
-				// contains a subset of the data.
-				//
-				// On Unix based system, the buffer will be truncated but no error will be returned.
-				length, addr, err := conn.ReadFrom(buffer)
-				if err != nil {
-					// don't log any deadline events.
-					e, ok := err.(net.Error)
-					if ok && e.Timeout() {
-						continue
-					}
-
-					// Closed network error string will never change in Go 1.X
-					// https://github.com/golang/go/issues/4373
-					opErr, ok := err.(*net.OpError)
-					if ok && strings.Contains(opErr.Err.Error(), "use of closed network connection") {
-						logger.Info("Connection has been closed")
-						return nil
-					}
-
-					logger.Errorf("Error reading from the socket %s", err)
-
-					// On Windows send the current buffer and mark it as truncated.
-					// The buffer will have content but length will return 0, addr will be nil.
-					if family == inputsource.FamilyUDP && isLargerThanBuffer(err) {
-						callback(buffer, inputsource.NetworkMetadata{RemoteAddr: addr, Truncated: true})
-						continue
-					}
-				}
-
-				if length > 0 {
-					callback(buffer[:length], inputsource.NetworkMetadata{RemoteAddr: addr})
-				}
-			}
-			return nil
-		})
-	}
-}
-
-func isLargerThanBuffer(err error) bool {
-	if runtime.GOOS != "windows" {
-		return false
-	}
-	return strings.Contains(err.Error(), windowErrBuffer)
 }
 
 func (l *Listener) Run(ctx context.Context) error {
